refactor(types): UnsentDanmakusRecord 从 exporter 移至 types/common
消除 engine 对 service 的反向依赖，类型定义归入 types 层。
