+++ src/danmaku_sender/ui/controllers/sender_controller.py
@@ -11,10 +11,10 @@ from danmaku_sender.core.engines.sender import DanmakuScheduler, DanmakuExecutor
 from danmaku_sender.core.engines.sender.delay_manager import DelayManager
 from danmaku_sender.core.entities.danmaku import Danmaku
 from danmaku_sender.core.types.result import DanmakuSendResult
-from danmaku_sender.core.types.common import VideoTarget
+from danmaku_sender.core.types.common import VideoTarget, UnsentDanmakusRecord
 from danmaku_sender.core.state import ApiAuthConfig, SenderConfig, AppState
 from danmaku_sender.core.services.danmaku_parser import DanmakuParser
-from danmaku_sender.core.services.danmaku_exporter import create_xml_from_danmakus, UnsentDanmakusRecord
+from danmaku_sender.core.services.danmaku_exporter import create_xml_from_danmakus
 from danmaku_sender.api.bili_api_client import BiliApiClient
 from danmaku_sender.utils.system_utils import KeepSystemAwake
