+++ src/danmaku_sender/ui/sender_page.py
@@ -17,8 +17,7 @@ from .controllers.sender_controller import SenderController
 
 from danmaku_sender.core.engines.sender.context import SendingContext
 from danmaku_sender.core.entities.video import VideoInfo
-from danmaku_sender.core.types.common import VideoTarget
-from danmaku_sender.core.services.danmaku_exporter import UnsentDanmakusRecord
+from danmaku_sender.core.types.common import VideoTarget, UnsentDanmakusRecord
 from danmaku_sender.core.state import AppState
 from danmaku_sender.utils.string_utils import parse_bilibili_link
 from danmaku_sender.utils.time_utils import format_duration
